requirements.txt
fastapi
uvicorn
groq
python-dotenv
httpx
python-multipart
